separates editor and game into hpp/cpp (now 3 compilation units crunch for almost all the lib???)
