+++ source/CMakeLists.txt
@@ -1,3 +1,3 @@
 cmake_minimum_required(VERSION 3.20)
 
-add_executable(App main.cpp)
+add_executable(App main.cpp editor.cpp game.cpp)
